+++ safe_ice/core/__init__.py
@@ -1,8 +1,15 @@
 """Core Safe-ICE algorithm and parameters."""
 
 from .adaptive_safe_ice import AdaptiveSafeICE
+from .ice_vmfnm import ICEvMFNM
 from .parameters import vMFNMParameters
 from .safe_ice import SafeICE
 from .safe_ice_optimized import OptimizedSafeICE
 
-__all__ = ["AdaptiveSafeICE", "OptimizedSafeICE", "SafeICE", "vMFNMParameters"]
+__all__ = [
+    "AdaptiveSafeICE",
+    "ICEvMFNM",
+    "OptimizedSafeICE",
+    "SafeICE",
+    "vMFNMParameters",
+]